+++ compiler/src/main/scala/dx/compiler/ApplicationCompiler.scala
@@ -1,6 +1,6 @@
 package dx.compiler
 
-import dx.api.{DxAccessLevel, DxApi, DxInstanceType, DxPath, DxUtils, InstanceTypeDB}
+import dx.api.{DxAccessLevel, DxApi, DxFileDescribe, DxPath, DxProject, DxUtils}
 import dx.core.Constants
 import dx.core.io.{DxWorkerPaths, StreamFiles}
 import dx.core.ir._
@@ -11,6 +11,7 @@ import dx.translator.ExtrasJsonProtocol._
 import dx.util.{CodecUtils, Logger}
 import spray.json._
 import wdlTools.generators.Renderer
+import scala.util.{Failure, Success, Try}
 
 object ApplicationCompiler {
   val DefaultAppletTimeoutInDays = 2
@@ -28,21 +29,27 @@ object ApplicationCompiler {
   private val AwsRegionKey = "region"
 }
 
-case class ApplicationCompiler(typeAliases: Map[String, Type],
-                               instanceTypeDb: InstanceTypeDB,
-                               runtimeAsset: Option[JsValue],
-                               runtimeJar: String,
-                               runtimePathConfig: DxWorkerPaths,
-                               runtimeTraceLevel: Int,
-                               separateOutputs: Boolean,
-                               streamFiles: StreamFiles.StreamFiles,
-                               waitOnUpload: Boolean,
-                               extras: Option[Extras],
-                               parameterLinkSerializer: ParameterLinkSerializer,
-                               useManifests: Boolean,
-                               dxApi: DxApi = DxApi.get,
-                               logger: Logger = Logger.get)
-    extends ExecutableCompiler(extras, parameterLinkSerializer, dxApi) {
+case class ApplicationCompiler(
+    typeAliases: Map[String, Type],
+    runtimeAsset: Option[JsValue],
+    runtimeJar: String,
+    runtimePathConfig: DxWorkerPaths,
+    runtimeTraceLevel: Int,
+    separateOutputs: Boolean,
+    streamFiles: StreamFiles.StreamFiles,
+    waitOnUpload: Boolean,
+    extras: Option[Extras],
+    parameterLinkSerializer: ParameterLinkSerializer,
+    useManifests: Boolean,
+    instanceTypeSelection: InstanceTypeSelection.InstanceTypeSelection,
+    defaultInstanceType: Option[String],
+    dxApi: DxApi = DxApi.get,
+    logger: Logger = Logger.get,
+    project: DxProject,
+    folder: String
+) extends ExecutableCompiler(extras, parameterLinkSerializer, dxApi) {
+  // database of available instance types for the user/org that owns the project
+  private lazy val instanceTypeDb = InstanceType.createDb(Some(project))
 
   // renderer for job script templates
   private lazy val renderer = Renderer()
@@ -142,9 +149,13 @@ case class ApplicationCompiler(typeAliases: Map[String, Type],
   }
 
   private def createRunSpec(applet: Application): (JsValue, Map[String, JsValue]) = {
-    val instanceType: DxInstanceType = applet.instanceType match {
-      case static: StaticInstanceType                => instanceTypeDb.apply(static.toInstanceTypeRequest)
-      case DefaultInstanceType | DynamicInstanceType => instanceTypeDb.defaultInstanceType
+    val instanceType: String = applet.instanceType match {
+      case static: StaticInstanceType =>
+        instanceTypeDb.apply(static.toInstanceTypeRequest).name
+      case DefaultInstanceType | DynamicInstanceType =>
+        // TODO: should we use the project default here rather than
+        //  picking one from the database?
+        defaultInstanceType.getOrElse(instanceTypeDb.defaultInstanceType.name)
     }
     // Generate the applet's job script
     val jobScript = generateJobScript(applet)
@@ -155,7 +166,7 @@ case class ApplicationCompiler(typeAliases: Map[String, Type],
         "systemRequirements" ->
           JsObject(
               "main" ->
-                JsObject("instanceType" -> JsString(instanceType.name))
+                JsObject("instanceType" -> JsString(instanceType))
           ),
         "distribution" -> JsString(Constants.OsDistribution),
         "release" -> JsString(Constants.OsRelease),
@@ -203,10 +214,44 @@ case class ApplicationCompiler(typeAliases: Map[String, Type],
         }
       case _ => Map.empty
     }
-    val bundledDepends = runtimeAsset match {
-      case Some(jsv) => Map("bundledDepends" -> JsArray(Vector(jsv)))
-      case None      => Map.empty
+
+    // Add platform Docker image dependency to bundledDepends
+    val bundledDependsDocker: Option[JsValue] = applet.container match {
+      case DxFileDockerImage(_, dxfile) => {
+        val id = dxfile.id
+
+        // If source project not specified for Docker image file, try to find it
+        val sourceProject = dxfile.project.getOrElse(
+            Try {
+              dxApi.getObject(id).describe() match {
+                case f: DxFileDescribe => dxApi.project(f.project)
+                case _                 => throw new RuntimeException(s"Expected ${id} to be a file")
+              }
+            } match {
+              case Success(project) => project
+              case Failure(ex)      => throw new RuntimeException(s"Unable to locate file ${id}", ex)
+            }
+        )
+
+        // If dependency on Docker image file in another project, clone
+        dxApi.cloneDataObject(id, sourceProject, project, folder)
+
+        val mapping = JsObject(
+            Constants.BundledDependsNameKey -> JsString(dxfile.describe().name),
+            Constants.BundledDependsIdKey -> JsObject(DxUtils.DxLinkKey -> JsString(dxfile.id)),
+            Constants.BundledDependsStagesKey -> JsArray(Vector.empty)
+        )
+        Some(mapping)
+      }
+      case _ => None
     }
+
+    // Include runtimeAsset in bundledDepends
+    val bundledDepends = Vector(runtimeAsset, bundledDependsDocker).flatten match {
+      case Vector()            => Map.empty
+      case bundledDependsItems => Map(Constants.BundledDependsKey -> JsArray(bundledDependsItems))
+    }
+
     val runSpec = JsObject(
         runSpecRequired ++ defaultTimeout ++ extrasOverrides ++ taskOverrides ++ taskSpecificOverrides ++ bundledDepends
     )
@@ -402,8 +447,7 @@ case class ApplicationCompiler(typeAliases: Map[String, Type],
     val fileDependencies = (
         applet.inputs
           .filter(param => param.dxType == Type.TFile)
-          .flatMap(fileDependenciesFromParam)
-          .toVector ++ applet.staticFileDependencies
+          .flatMap(fileDependenciesFromParam) ++ applet.staticFileDependencies
     ).distinct
     val fileDependenciesDetails = Option
       .when(fileDependencies.nonEmpty)(
@@ -490,21 +534,24 @@ case class ApplicationCompiler(typeAliases: Map[String, Type],
       }
     // compress and base64 encode the source code
     val sourceEncoded = CodecUtils.gzipAndBase64Encode(applet.document.toString)
-    // serialize the pricing model, and make the prices opaque.
-    val dbOpaque = InstanceTypeDB.opaquePrices(instanceTypeDb)
-    val dbOpaqueEncoded = CodecUtils.gzipAndBase64Encode(dbOpaque.toJson.prettyPrint)
+    // compress and base64 encode the instance types, unless we specify that we want to
+    // resolve them at runtime, which requires that the user running the applet has
+    // permission to describe the project it is running in
+    val dbEncoded = Option.when(instanceTypeSelection == InstanceTypeSelection.Static) {
+      CodecUtils.gzipAndBase64Encode(instanceTypeDb.toJson.prettyPrint)
+    }
     // serilize default runtime attributes
     val defaultRuntimeAttributes: JsValue = extras
       .flatMap(ex =>
         ex.defaultRuntimeAttributes.map(attr => JsObject(ValueSerde.serializeMap(attr)))
       )
       .getOrElse(JsNull)
-    val auxDetails = Map(
-        Constants.SourceCode -> JsString(sourceEncoded),
-        Constants.ParseOptions -> applet.document.optionsToJson,
-        Constants.InstanceTypeDb -> JsString(dbOpaqueEncoded),
-        Constants.RuntimeAttributes -> defaultRuntimeAttributes
-    )
+    val auxDetails = Vector(
+        Some(Constants.SourceCode -> JsString(sourceEncoded)),
+        Some(Constants.ParseOptions -> applet.document.optionsToJson),
+        dbEncoded.map(db => Constants.InstanceTypeDb -> JsString(db)),
+        Some(Constants.RuntimeAttributes -> defaultRuntimeAttributes)
+    ).flatten.toMap
     val useManifestsDetails = if (useManifests) {
       Map(Constants.UseManifests -> JsBoolean(true))
     } else {