+++ scripts/run_tests.py
@@ -6,6 +6,7 @@ import glob
 import hashlib
 import json
 import os
+import random
 import re
 import sys
 import subprocess
@@ -721,13 +722,18 @@ def build_test(tname, project, folder, version_id, compiler_flags):
     desc = test_files[tname]
     print("build {} {}".format(desc.kind, desc.name))
     print("Compiling {} to a {}".format(desc.source_file, desc.kind))
+    # both static and dynamic instance type selection should work,
+    # so we can test them at random
+    instance_type_selection = random.choice(["static", "dynamic"])
     cmdline = [ "java", "-jar",
                 os.path.join(top_dir, "dxCompiler-{}.jar".format(version_id)),
                 "compile",
                 desc.source_file,
                 "-force",
                 "-folder", folder,
-                "-project", project.get_id() ]
+                "-project", project.get_id(),
+                "-instanceTypeSelection", instance_type_selection
+                ]
     if "manifest" in desc.source_file:
         cmdline.append("-useManifests")
     cmdline += compiler_flags